requirements.txt
selenium
requests
flask
python-dotenv
webdriver-manager
google-generativeai==0.1.0rc1
tkinter
Pillow>=10.0.0
